Move TokensExtractor to NodeExtractor
